Update based on master + respond to some comments
