+++ kernel/kernel-defaults/src/test/scala/io/delta/kernel/defaults/DeltaTableWritesTransactionBuilderV2Suite.scala
@@ -21,6 +21,7 @@ import scala.collection.immutable.Seq
 import io.delta.kernel.{Operation, TableManager, Transaction}
 import io.delta.kernel.commit.{CommitMetadata, CommitResponse, Committer}
 import io.delta.kernel.data.Row
+import io.delta.kernel.defaults.utils.WriteUtilsWithV2Builders
 import io.delta.kernel.engine.Engine
 import io.delta.kernel.exceptions.{KernelException, TableAlreadyExistsException}
 import io.delta.kernel.expressions.Column
@@ -32,86 +33,7 @@ import io.delta.kernel.types.StructType
 import io.delta.kernel.utils.CloseableIterable.emptyIterable
 import io.delta.kernel.utils.CloseableIterator
 
-trait UseTransactionBuilderV2 extends DeltaTableWriteSuiteBase {
-
-  // scalastyle:off argcount
-  override def createTxn(
-      engine: Engine = defaultEngine,
-      tablePath: String,
-      isNewTable: Boolean = false,
-      schema: StructType = null,
-      partCols: Seq[String] = null,
-      tableProperties: Map[String, String] = null,
-      clock: Clock = () => System.currentTimeMillis,
-      withDomainMetadataSupported: Boolean = false,
-      maxRetries: Int = -1,
-      clusteringColsOpt: Option[List[Column]] = None,
-      logCompactionInterval: Int = 10,
-      txnId: Option[(String, Long)] = None,
-      tablePropertiesRemoved: Set[String] = null): Transaction = {
-    // scalastyle:on argcount
-    def tblPropertiesWithDomainMetadata = {
-      val origTblProps = if (tableProperties != null) tableProperties else Map()
-      val dmTblProps = if (withDomainMetadataSupported) {
-        Map(TableFeatures.SET_TABLE_FEATURE_SUPPORTED_PREFIX + "domainMetadata" -> "supported")
-      } else {
-        Map()
-      }
-      origTblProps ++ dmTblProps
-    }
-
-    if (isNewTable) {
-      var txnBuilder = TableManager.buildCreateTableTransaction(
-        tablePath,
-        schema,
-        testEngineInfo)
-      if (partCols != null) {
-        txnBuilder = txnBuilder.withDataLayoutSpec(
-          DataLayoutSpec.partitioned(partCols.map(new Column(_)).asJava))
-      }
-      if (tableProperties != null || withDomainMetadataSupported) {
-        txnBuilder = txnBuilder.withTableProperties(tblPropertiesWithDomainMetadata.asJava)
-      }
-      if (clusteringColsOpt.nonEmpty) {
-        txnBuilder = txnBuilder.withDataLayoutSpec(
-          DataLayoutSpec.clustered(clusteringColsOpt.get.asJava))
-      }
-      if (maxRetries >= 0) {
-        txnBuilder = txnBuilder.withMaxRetries(maxRetries)
-      }
-      txnBuilder.build(engine)
-    } else {
-      var txnBuilder = TableManager.loadSnapshot(tablePath)
-        .asInstanceOf[SnapshotBuilderImpl].build(engine)
-        .buildUpdateTableTransaction(testEngineInfo, Operation.WRITE)
-      /*
-      TODO: in a separate PR clean up usage of createTxn such that schema != null ONLY when it
-        should either be updated for schema evolution or for a new table
-      if (schema != null) {
-        txnBuilder = txnBuilder.withUpdatedSchema(schema)
-      }
-       */
-      if (clusteringColsOpt.nonEmpty) {
-        txnBuilder = txnBuilder.withClusteringColumns(clusteringColsOpt.get.asJava)
-      }
-      if (tableProperties != null || withDomainMetadataSupported) {
-        txnBuilder = txnBuilder.withTablePropertiesAdded(tblPropertiesWithDomainMetadata.asJava)
-      }
-      if (maxRetries >= 0) {
-        txnBuilder = txnBuilder.withMaxRetries(maxRetries)
-      }
-      txnBuilder = txnBuilder.withLogCompactionInterval(logCompactionInterval)
-      txnId.foreach { case (appId, txnVer) =>
-        txnBuilder = txnBuilder.withTransactionId(appId, txnVer)
-      }
-      if (tablePropertiesRemoved != null) {
-        txnBuilder = txnBuilder.withTablePropertiesRemoved(tablePropertiesRemoved.asJava)
-      }
-      txnBuilder.build(engine)
-    }
-  }
-}
-
+// TODO update docs + new tests
 /**
  * Tests for the V2 transaction builders [[CreateTableTransactionBuilder]] and
  * [[UpdateTableTransactionBuilder]]. We don't cover the full scope of tests we have for
@@ -126,7 +48,11 @@ trait UseTransactionBuilderV2 extends DeltaTableWriteSuiteBase {
  * <p>In the future, when we deprecate the V1 builder, we should switch all the existing suites
  * to use the V2 builders -- and maintain something similar to this for the V1 builder.
  */
-class TransactionBuildersV2Suite extends DeltaTableWritesSuite with UseTransactionBuilderV2 {
+class DeltaTableWritesTransactionBuilderV2Suite extends DeltaTableWritesSuite
+    with WriteUtilsWithV2Builders {
+
+  // TablePropertiesTransactionBuilderV2Suite tests table property validation, normalization and
+  // unset/set overlap for Create + Update
 
   // Tested in DeltaTableWritesSuite: setTxnOpt (covered by idempotent writes test)
   // Tested in DeltaTableWritesSuite: validateKernelCanWriteToTable (covered by unsupported
@@ -231,8 +157,3 @@ class TransactionBuildersV2Suite extends DeltaTableWritesSuite with UseTransacti
     }
   }
 }
-
-// This suite tests table property validation, normalization, and unset/set overlap for
-// Create + Update
-class TablePropertiesTransactionBuilderV2Suite extends TablePropertiesSuiteBase
-    with UseTransactionBuilderV2 {}
